new route for contactForm and test skipped
